feat(sidebar): refactor sidebar with animations and new navigation structure
- Add framer-motion for animated transitions
- Reorganize navigation items with dividers
- Improve mobile and desktop interaction
- Rename component from ApolloSidebar to AppSidebar
