+++ package.json
@@ -49,6 +49,7 @@
     "cmdk": "1.0.4",
     "date-fns": "4.1.0",
     "embla-carousel-react": "8.5.1",
+    "framer-motion": "^12.23.15",
     "geist": "latest",
     "input-otp": "1.4.1",
     "lucide-react": "^0.454.0",
